Turned original app into legacy file, refactored it into smoother, easier to read code
